server/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
decorator==5.1.1
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==3.18.0
future==1.0.0
geocoder==1.38.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==1.6.1
python-dotenv==1.0.1
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
six==1.16.0
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.1
